Move Auth and Login classes to utils #59
