+++ backend/tests/services/test_deploy_container_service.py
@@ -4,7 +4,7 @@ import unittest
 from src.app import app
 from src.classes.customer import Customer
 
-from tests.services.login import TestingLogin
+from tests.utils.login import TestingLogin
 
 from config.server_environment import TESTING_DATABASE
