+++ backend/tests/utils/auth.py
@@ -1,0 +1,4 @@
+class Auth:
+    def __init__(self, u, p):
+        self.username = u
+        self.password = p
